pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dotai"
version = "0.1.0"
description = "One config for all your AI coding tools. Define rules once, sync to Claude, Cursor, Copilot, Windsurf, Cline, and more."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Lakshmi Sravya Vedantham" },
]
keywords = [
    "ai",
    "claude",
    "cursor",
    "copilot",
    "windsurf",
    "cline",
    "aider",
    "developer-tools",
    "configuration",
    "dotfiles",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
dotai = "dotai.cli:app"

[project.urls]
Homepage = "https://github.com/LakshmiSravyaVedantham/dotai"
Repository = "https://github.com/LakshmiSravyaVedantham/dotai"
Issues = "https://github.com/LakshmiSravyaVedantham/dotai/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "black>=23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dotai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=dotai --cov-report=term-missing"

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
target-version = ["py39"]
line-length = 88
